requirements.txt
aiogram==2.25.1
tortoise-orm==0.19.0
aerich==0.10.0
pydantic==1.10.7
apscheduler==3.9.1
python-dotenv==1.0.0
aiohttp==3.8.1 